Invoke handler functions from apply function
 - Remove ConsensourceTransactionHandler functions
 - Move handler tests to respective handler function files
 - Create a test module for MockTransactionContext and helper test functions
