+++ src/transaction_handler/mod.rs
@@ -5,3 +5,4 @@ pub mod factory;
 pub mod handler;
 pub mod organization;
 pub mod standard;
+pub mod test_utils;
